refactor: move redux to vendor folder
