+++ packages/analytics-core/src/utils/debug.js
@@ -1,4 +1,4 @@
-import { compose } from '../redux'
+import { compose } from '../vendor/redux'
 import globalContext from './global'
 
 export function Debug() {
